refactor: ProviderCard css to modular css
