+++ src/pages/ProviderListPage/ProviderList.tsx
@@ -1,6 +1,6 @@
 import { Fragment, useContext, useState } from "react";
 import { ProviderContext } from "../../contexts/providerContext";
-import ProviderCard from "../../components/ProviderCard";
+import ProviderCard from "../../components/ProviderCard/ProviderCard";
 import Loader from "../../components/Loader";
 import styles from "./ProviderList.module.css"
